# openapi.yaml
openapi: 3.0.3
info:
  title: Sunset Beach Resort & Spa — Admin/Booking API
  description: >
    Contract reverse-engineered from the current Next.js route handlers under
    `app/api/**/route.ts`. It documents exactly what each route accepts and
    returns today (including quirks such as Prisma `Decimal`/`Date` JSON
    serialization), so it can serve as a migration contract for a backend
    rewrite in another language. No behavior was changed to produce this spec.
  version: "1.0.0"
  license:
    name: Proprietary
  x-codegen-notes: >
    A property marked `x-sensitive: true` (guest email/phone, payment notes) is redacted as
    "[REDACTED]" in its generated model's toString() by the customized pojo.mustache template
    under openapi-templates/ - see scripts/generate-api.sh, which passes --template-dir. This is
    the one and only place PII redaction needs to be declared; it survives regeneration
    automatically. Do not hand-edit a generated model's toString() to redact a field instead -
    mark the field here and regenerate.
servers:
  - url: /api
    description: Relative to the deployed Next.js app origin.
...
security:
  - bearerAuth: []
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [Rooms]
      summary: List all rooms
      description: >
        Requires CASHIER or above - a CASHIER quoting a room type for a walk-in via
        `POST /bookings/staff` needs to read this list, not just create against it (the same
        "action allowed, prerequisite read blocked" fix already applied to RoomUnits/Availability).
        Writes (create/update/delete, images) stay MANAGER-only. No pagination (small hotel).
        Broader context: at this property CASHIER is a combined cashier/front-desk role (see
        `POST /bookings/staff`'s description) - this is a deliberate scoping decision covering
        this operation and `GET /pricing/{roomId}`, `GET /bookings`, `GET /bookings/calendar`,
        and `GET /availability/{roomId}` together, not five independent oversights to flag one
        by one in review.
      responses:
        "200":
          description: All rooms.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createRoom
      tags: [Rooms]
      summary: Create a room
      description: Requires any authenticated staff session (ADMIN or MANAGER).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          description: Body failed `roomSchema` validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rooms/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getRoom
      tags: [Rooms]
      summary: Get a room by id
      description: >
        Requires CASHIER or above (see `GET /rooms`). Also used to prefill the admin edit form
        for MANAGER/ADMIN.
      responses:
        "200":
          description: The room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Room not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
    patch:
      operationId: updateRoom
      tags: [Rooms]
      summary: Update a room
      description: Requires any authenticated staff session (ADMIN or MANAGER). Full replacement of all `roomSchema` fields (no partial update).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "200":
          description: Updated room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          description: Body failed `roomSchema` validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Room not found (Prisma `P2025`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
    delete:
      operationId: deleteRoom
      tags: [Rooms]
      summary: Delete a room
      description: Requires any authenticated staff session (ADMIN or MANAGER).
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Room not found (Prisma `P2025`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
        "409":
          description: Room has existing bookings or room units and cannot be deleted (Prisma `P2003`, FK constraint).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: This room has existing bookings and can't be deleted.
  /rooms/{id}/images:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: uploadRoomImages
      tags: [Rooms]
      summary: Upload one or more room images
      description: >
        Requires any authenticated staff session (ADMIN or MANAGER). Files
        are written to `public/uploads/rooms/{id}/` under a randomized
        filename and appended to `Room.images`. All files in the request are
        validated (type + size) before any are written; the first violation
        aborts the whole request.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more image files. Field name must be repeated as `files` per file.
              required: [files]
      responses:
        "201":
          description: Room updated with the new image paths appended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          description: No files provided, an unsupported MIME type, or a file over 8MB.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                noFiles:
                  value: { error: "No files provided" }
                unsupportedType:
                  value: { error: "Unsupported file type: image/gif" }
                tooLarge:
                  value: { error: "photo.jpg exceeds the 8MB limit" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Room not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
    delete:
      operationId: deleteRoomImage
      tags: [Rooms]
      summary: Remove a room image
      description: >
        Requires any authenticated staff session (ADMIN or MANAGER). Removes
        the given path from `Room.images` and, if it points under this
        room's upload directory, deletes the file from disk (best-effort —
        failure to unlink is ignored).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Exact image path as stored in `Room.images`, e.g. `/uploads/rooms/{id}/169...-ab12cd.jpg`.
      responses:
        "200":
          description: Room updated with the image path removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          description: '`path` missing or not a string.'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: path is required
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Room not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT returned by `POST /auth/login`, sent as `Authorization: Bearer
        <token>`. Contains `sub` (user id), `email`, `role`, standard `iat`/`exp`
        (default TTL 7 days, configurable via `JWT_TTL_DAYS`). Every route in
        this file except `POST /bookings` and the unauthenticated `/auth/*`
        endpoints (`login`, `logout`) requires a valid token, returning 401 if
        missing/invalid/expired, and 403 where the route additionally requires
        role `ADMIN`.
  responses:
    Unauthorized:
      description: No valid JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            error: Unauthorized
...
  schemas:
...
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
      description: 'Generic error shape used by all hand-written `{ error: "..." }` responses.'
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          description: Output of Zod's `error.flatten()`.
          properties:
            formErrors:
              type: array
              items:
                type: string
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
      example:
        error:
          formErrors: []
          fieldErrors:
            guestEmail: ["Invalid email"]
    OkTrue:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
...
    Room:
      type: object
      description: >
        Despite the name this is a **room type**, not a single physical room — individual
        sellable units are separate `RoomUnit` rows (see the RoomUnits tag). `activeUnitCount`
        is computed from those rows (`count(RoomUnit WHERE roomId = id AND isActive)`), not
        stored. `basePrice` is a Prisma `Decimal` — it serializes to a **string**, not a JSON
        number. `createdAt` serializes as an ISO-8601 datetime string.
      required: [id, name, description, capacity, activeUnitCount, basePrice, images, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        capacity:
          type: integer
        activeUnitCount:
          type: integer
          minimum: 0
          description: 'Computed - number of active (`isActive: true`) `RoomUnit` rows of this type.'
        basePrice:
          type: string
          description: Decimal(10,2) rendered as a string, e.g. `"1500.00"`.
        images:
          type: array
          items:
            type: string
          description: Paths under `/uploads/rooms/{id}/...`.
        createdAt:
          type: string
          format: date-time
    RoomInput:
      type: object
      description: >
        Body of `roomSchema`. Full replacement on PATCH — no partial update. Does not include a
        unit count - a room type has zero sellable units until `RoomUnit`s are added for it via
        `POST /room-units`.
      required: [name, description, capacity, basePrice]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        description:
          type: string
          minLength: 10
        capacity:
          type: integer
          minimum: 1
          maximum: 20
          description: Coerced to a number server-side (`z.coerce.number()`); a numeric string is also accepted.
        basePrice:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Coerced to a number server-side (`z.coerce.number()`); a numeric string is also accepted.
...