# src/main/java/com/sunsetbeach/api/RoomsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sunsetbeach.api;

import com.sunsetbeach.model.DeleteRoomImageRequest;
import com.sunsetbeach.model.ErrorMessage;
import com.sunsetbeach.model.OkTrue;
import com.sunsetbeach.model.Room;
import com.sunsetbeach.model.RoomInput;
import com.sunsetbeach.model.ValidationError;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
public interface RoomsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /rooms : Create a room
     * Requires any authenticated staff session (ADMIN or MANAGER).
     *
     * @param roomInput  (required)
     * @return Room created. (status code 201)
     *         or Body failed &#x60;roomSchema&#x60; validation. (status code 400)
     *         or No valid JWT. (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/rooms",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Room> createRoom(
         @Valid @RequestBody RoomInput roomInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"formErrors\" : [ ], \"fieldErrors\" : { \"guestEmail\" : [ \"Invalid email\" ] } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /rooms/{id} : Delete a room
     * Requires any authenticated staff session (ADMIN or MANAGER).
     *
     * @param id  (required)
     * @return Deleted. (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Room not found (Prisma &#x60;P2025&#x60;). (status code 404)
     *         or Room has existing bookings or room units and cannot be deleted (Prisma &#x60;P2003&#x60;, FK constraint). (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/rooms/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<OkTrue> deleteRoom(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"ok\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /rooms/{id}/images : Remove a room image
     * Requires any authenticated staff session (ADMIN or MANAGER). Removes the given path from &#x60;Room.images&#x60; and, if it points under this room&#39;s upload directory, deletes the file from disk (best-effort — failure to unlink is ignored). 
     *
     * @param id  (required)
     * @param deleteRoomImageRequest  (required)
     * @return Room updated with the image path removed. (status code 200)
     *         or &#x60;path&#x60; missing or not a string. (status code 400)
     *         or No valid JWT. (status code 401)
     *         or Room not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/rooms/{id}/images",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Room> deleteRoomImage(
         @PathVariable("id") String id,
         @Valid @RequestBody DeleteRoomImageRequest deleteRoomImageRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /rooms/{id} : Get a room by id
     * Requires CASHIER or above (see &#x60;GET /rooms&#x60;). Also used to prefill the admin edit form for MANAGER/ADMIN. 
     *
     * @param id  (required)
     * @return The room. (status code 200)
     *         or No valid JWT. (status code 401)
     *         or Room not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/rooms/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Room> getRoom(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /rooms : List all rooms
     * Requires CASHIER or above - a CASHIER quoting a room type for a walk-in via &#x60;POST /bookings/staff&#x60; needs to read this list, not just create against it (the same \&quot;action allowed, prerequisite read blocked\&quot; fix already applied to RoomUnits/Availability). Writes (create/update/delete, images) stay MANAGER-only. No pagination (small hotel). Broader context: at this property CASHIER is a combined cashier/front-desk role (see &#x60;POST /bookings/staff&#x60;&#39;s description) - this is a deliberate scoping decision covering this operation and &#x60;GET /pricing/{roomId}&#x60;, &#x60;GET /bookings&#x60;, &#x60;GET /bookings/calendar&#x60;, and &#x60;GET /availability/{roomId}&#x60; together, not five independent oversights to flag one by one in review. 
     *
     * @return All rooms. (status code 200)
     *         or No valid JWT. (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/rooms",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Room>> listRooms(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /rooms/{id} : Update a room
     * Requires any authenticated staff session (ADMIN or MANAGER). Full replacement of all &#x60;roomSchema&#x60; fields (no partial update).
     *
     * @param id  (required)
     * @param roomInput  (required)
     * @return Updated room. (status code 200)
     *         or Body failed &#x60;roomSchema&#x60; validation. (status code 400)
     *         or No valid JWT. (status code 401)
     *         or Room not found (Prisma &#x60;P2025&#x60;). (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/rooms/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Room> updateRoom(
         @PathVariable("id") String id,
         @Valid @RequestBody RoomInput roomInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"formErrors\" : [ ], \"fieldErrors\" : { \"guestEmail\" : [ \"Invalid email\" ] } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /rooms/{id}/images : Upload one or more room images
     * Requires any authenticated staff session (ADMIN or MANAGER). Files are written to &#x60;public/uploads/rooms/{id}/&#x60; under a randomized filename and appended to &#x60;Room.images&#x60;. All files in the request are validated (type + size) before any are written; the first violation aborts the whole request. 
     *
     * @param id  (required)
     * @param files One or more image files. Field name must be repeated as &#x60;files&#x60; per file. (required)
     * @return Room updated with the new image paths appended. (status code 201)
     *         or No files provided, an unsupported MIME type, or a file over 8MB. (status code 400)
     *         or No valid JWT. (status code 401)
     *         or Room not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/rooms/{id}/images",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    default ResponseEntity<Room> uploadRoomImages(
         @PathVariable("id") String id,
         @RequestPart(value = "files", required = true) List<MultipartFile> files
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}